requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlib
colorama
streamlit
playsound==1.2.2
speechrecognition==3.14.0
#pyaudio